% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ans.R
\name{ans}
\alias{ans}
\title{Adjusted Net Savings (ANS)}
\usage{
ans(gross_savings, edu_expenditure, resource_depletion, pollution_damage)
}
\arguments{
\item{gross_savings}{Numeric gross savings.}

\item{edu_expenditure}{Numeric education expenditure.}

\item{resource_depletion}{Numeric natural resource depletion.}

\item{pollution_damage}{Numeric pollution damages.}
}
\value{
Numeric adjusted net savings.
}
\description{
Computes a simple Adjusted Net Savings proxy:
gross savings + education expenditure - resource depletion - pollution damages.
}
\examples{
ans(gross_savings = 100, edu_expenditure = 10, resource_depletion = 15, pollution_damage = 5)
}
