% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_jacobian.R
\name{plot_sensitivity_heatmap}
\alias{plot_sensitivity_heatmap}
\title{Plot a sensitivity heatmap}
\usage{plot_sensitivity_heatmap(sens_df)}
\arguments{\item{sens_df}{Output of \code{sensitivity_jacobian()}.}}
\value{A ggplot object.}
\description{Creates a heatmap from a sensitivity/Jacobian table.}
\examples{
times <- seq(0, 5, by = 1)
x0 <- c(K = 1, H = 1, N = 1, C = 0, P = 1, A = 1)
run <- catalyst_run(times, x0, include_phase_plane = FALSE, include_sensitivity = TRUE)
plot_sensitivity_heatmap(run$sensitivities)
}
