% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_catalyst_run.R
\name{print.catalyst_run}
\alias{print.catalyst_run}
\title{Print a Catalyst run (friendly summary)}
\usage{
\method{print}{catalyst_run}(x, ...)
}
\arguments{
\item{x}{A \code{catalyst_run} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
S3 method for printing a \code{catalyst_run} object.
}
\examples{
x <- catalyst_demo()
print(x)
}
