% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_catalyst.R
\name{theme_catalyst}
\alias{theme_catalyst}
\title{A Minimal Plot Theme for Catalyst Analytics R}
\usage{
theme_catalyst(base_size = 12)
}
\arguments{
\item{base_size}{Base font size for the theme.}
}
\value{
A ggplot2 theme object.
}
\description{
A Minimal Plot Theme for Catalyst Analytics R
}
\examples{
# Use in a ggplot
ggplot2::ggplot(mtcars, ggplot2::aes(mpg, wt)) +
  ggplot2::geom_point() +
  theme_catalyst()
}
